feat: basic ship162 plugin.
unlike jet1090 (which writes to redis timeseries to persist trajectory), we do
not simply forward the data to the frontend, keeping track of an in-memory store

this is because in the future we will write our own persistent data store for
history recording and playback.

the bbox filtering has also been isolated into the tangram core and shared
across the two.

also made the cursor pointy when hovering over an entity